requirements.txt
requests
beautifulsoup4
feedparser
python-docx
playwright
Flask
Werkzeug
gunicorn